created a folder for singleMovie. Editted the Navbar to make it more organized and created a Cart Component
