+++ client/app/AppRoutes.js
@@ -8,7 +8,7 @@ import Movies from '../features/movies/Movies'
 import Home from '../features/home/Home'
 import Contact from '../features/contact/Contact'
 import Shop from '../features/shop/Shop'
-import SingleMovie from '../features/SingleMovie'
+import SingleMovie from '../features/singleMovie/SingleMovie'
 
 /**
  * COMPONENT
@@ -26,7 +26,7 @@ const AppRoutes = () => {
         <div>
             {isLoggedIn ? (
                 <Routes>
-                  <Route path="/home" element={<Home />}></Route>
+                    <Route path="/home" element={<Home />}></Route>
                     <Route path="/contact" element={<Contact />}></Route>
                     <Route path="/shop" element={<Shop />} />
                     <Route path={`/movies/:id`} element={<SingleMovie />} />
