+++ public/style.css
@@ -1,13 +1,5 @@
-body {
-    font-family: sans-serif;
-}
-
 a {
     text-decoration: none;
-}
-
-label {
-    display: block;
 }
 
 nav a {
@@ -20,11 +12,26 @@ form div {
 }
 
 .logo {
-    width: 600px;
+    width: 150px;
+    align-items: center;
+}
+
+.navbar {
+    display: flex;
+}
+
+.navbar-container {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    font-size: 20px;
+}
+
+.nav-right {
+    /* display: flex; */
+    justify-content: center;
+    border: black 1px solid;
     margin-left: auto;
-    margin-right: auto;
-    display: block;
-    margin-bottom: 20px;
 }
 
 @font-face {
@@ -79,33 +86,25 @@ form div {
     padding: 5px;
 }
 
-#navbar {
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  border: 1px solid black;
+#homepage-about {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-indent: 30px;
 }
 
-#homepage-about{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-indent: 30px;
+.about-title {
+    font-family: 'ITC Machine';
+    text-align: center;
+    margin-right: 50px;
 }
 
-.about-title{
-  font-family: 'ITC Machine';
-  text-align: center;
-  margin-right: 50px;
+.homepage-paragraphs {
+    max-width: 650px;
 }
 
-.homepage-paragraphs{
-  max-width: 650px;
+.home-shop {
+    font-family: 'ITC Machine';
+    display: flex;
+    justify-content: center;
 }
-
-.home-shop{
-  font-family: 'ITC Machine';
-  display: flex;
-  justify-content: center;
-}
-
